+++ src/server/mod.rs
@@ -6,7 +6,8 @@ mod service;
 pub use self::error::{HandshakeError, ServerError};
 pub use self::handshake::{Handshake, HandshakeAck, HandshakeAmqp, HandshakeAmqpOpened};
 pub use self::sasl::Sasl;
-pub use self::service::Server;
+pub use self::service::{Server, ServerBuilder};
+
 pub use crate::codec::protocol::Transfer;
 pub use crate::control::{ControlFrame, ControlFrameKind};
 pub use crate::error::{Error, LinkError};